Move BasicAuthenticator into server-common auth package
