+++ authenticators/authenticator-basic/build.gradle.kts
@@ -24,11 +24,6 @@ plugins {
 }
 
 dependencies {
-  implementation(project(":api"))
-  implementation(project(":common"))
-  implementation(project(":core"))
-  implementation(project(":server-common"))
-
   testImplementation(libs.junit.jupiter.api)
   testRuntimeOnly(libs.junit.jupiter.engine)
 }
